ai_newsletter/requirements.txt
anthropic>=0.40.0
feedparser>=6.0.0
requests>=2.31.0
jinja2>=3.1.0
python-dateutil>=2.8.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
